Setup of the DevEnvironment after the great purge. Implemented a perfectly working ScrollBar, yet it the scissoring of the ScollableArea currently is borked when it is placed inside a other component that already scissors. A update to the StandardRenderManager is required to make it work properly.
Added:
  + Dev files generated by Forge
  + ComponentContentArea: A simple wrapper component that allows easy handling of its subcomponents. Very usefull for situations where one components value determines the position of a group of others.
  + IContentAreaHost: Components that use a ComponentContentArea need to implement this interface to allow the ContentArea to retrieve its child elements. Important here is that the call to the registerContentComponents method provides the ContentArea which requests its contents and that the registerNewComponent method should be called on the Area, not as usual by component registration methods on the implementer of this interface. On top of this you will need to pass the added component the ContentArea as its parent instead of (again as usual) the class implementing this interface.
  + ComponentScrollableArea: Completely working implementation of a ScrollableArea. To track the changes to the ScrollBar in the UI, handle the ScrollBarValueChangedEvent.
  + ComponentScrollBar: Added the buttons to work properly and a Plane to be used as a detection for clicks in the InnerArea where no Button currently is. Also added the Animation component when clicked in the InnerArea to order the ScrollBar to that position.

Fixed:
  * ComponentButton: Fixed the content filter crash using OR's instead of AND's what it should be.
  * CoreComponent: Fixed the CoreComponent returning the root as its parent.
  * CoreLedger: Fixed the CoreLedgers registration behaviour for IGUIBasedComponentHost implementing Components.
  * InformationLedger: Now allows Scrolling, and has a properly fixed size.
  * StandardRenderManager: Fixed an NPE caused by (again) using OR's instead of AND's.
  * ScrollBarComponentState: Fixed the Scrolling and implemented the go to position animation.
  * ButtonInputEventHandler: Fixed an NPE caused by (again) using OR's instead of AND's.
  * MouseManager: Switched over to using the ClientTickEvent to retrieve the Mouse position, now directly grabs it from Mouse.
  * CoreClientProxy: Added the ButtonInputEventHandler to the standardly registered EventHandlers.
  * Colors: Some colors where not upgraded into MinecraftColors properly.
  * Textures: Fixed the texture positions of the ScrollBarButton icons.

+++ src/com/smithsmodding/smithscore/client/events/gui/ScrollBarValueChangedEvent.java
@@ -11,7 +11,7 @@ public class ScrollBarValueChangedEvent extends GuiInputEvent {
     int maximum;
     int minimum;
 
-    public ScrollBarValueChangedEvent (String componentID, int value, int maximum, int minimum) {
+    public ScrollBarValueChangedEvent (String componentID, float value, int maximum, int minimum) {
         super(InputTypes.SCROLLED, componentID, String.valueOf(value));
 
         this.maximum = maximum;
